Implement seat-based gating of display present paths across all drivers
- Added `SeatGate` abstraction to enforce live lease checks for display present/flip operations.
- Updated `SeatRegistry` to mint and verify seat leases, ensuring revoked clients are refused explicitly (`DriverError::SeatRevoked`).
- Integrated seat gating into software (`present`) and hardware (`present_layers`) paths for `rpi_hvs`, `vesa`, and `framebuffer` drivers.
- Documented changes in `plans/DISPLAY.md`, `docs/src/drivers/display.md`, and various driver READMEs.
- Verified gating behavior with new unit tests, including revoked lease refusal and fresh foreground lease acceptance.
- Updated `DISPLAY.md` D4 progress to reflect completion.

+++ include/rustos/rustos_syscall.h
@@ -165,7 +165,7 @@ uint64_t ros_sys_users_db_read(void * a0, uintptr_t a1);
 uint64_t ros_sys_console_count(void);
 int32_t ros_sys_stream_input_mode(uint32_t a0, uint32_t a1);
 uint64_t ros_sys_key_inject(void * a0, uintptr_t a1);
-int32_t ros_sys_display_acquire(void);
+uint64_t ros_sys_display_acquire(void);
 int32_t ros_sys_display_release(void);
 uint64_t ros_sys_keyboard_read(void * a0, uintptr_t a1);
 uint64_t ros_sys_mmio_map(uint64_t a0, uintptr_t a1, uintptr_t a2);
